requirements.txt
requests
paho-mqtt
psutil
argparse
oyaml
redis
